third_party/WebKit/LayoutTests/external/wpt/cssom-view/cssom-getClientRects-002-expected.txt

This is a testharness.js-based test.
FAIL CSSOM View - GetClientRects().length is the same regardless source new linesassert_equals: count2 expected 1 but got 2
Harness: the test ran to completion.
